requirements.txt
boto3==1.40.59
aws-cdk-lib==2.221.0
constructs>=10.0.0,<11.0.0
aws-lambda-powertools==3.22.0
cdk-s3-vectors==0.1.6
boto3==1.40.59
fastmcp==2.13.0.2
mcp==1.19.0